requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
openpyxl>=3.1.2
chromadb>=0.4.22
google-genai>=0.3.0
pydantic>=2.5.0
python-dotenv>=1.0.0